make position input text fields float compatible

+++ Assets/Scripts/UI/Rakel/RakelPositionZController.cs
@@ -8,7 +8,7 @@ public class RakelPositionZController : InputFieldController
 
     override public void OnValueChanged(string arg0)
     {
-        int value = int.Parse(arg0);
+        float value = float.Parse(arg0);
         OilPaintEngine.UpdateRakelPositionZ(value);
     }
 }
